requirements.txt
gallery-dl==1.23.5
Pillow==9.3.0
requests==2.28.1
